+++ Curves/DTW.hpp
@@ -5,4 +5,4 @@
 
 
 double getValueDTW(vector<Curve *>::iterator queryCurve,vector<Curve *>::iterator inputCurve);
-double distance(int x1, int x2, int y1, int y2);
+double distance(float x1, float x2, float y1, float y2);
